地形生成処理と生成された地形から異動経路を探索するExplorerを作成

+++ Assets/Common/PixelTerrain/Examples/Scripts/DropItem.cs
@@ -12,12 +12,12 @@ namespace Common.PixelTerrain {
 		[SerializeField, Range(1f, 20f)]
 		private float _speed = 10f;
 
-		private byte _id;
+		private int _id;
 		private string _itemName;
 		private SpriteRenderer _sprite;
 		private ItemCollector _targetCollector;
 
-		public byte id {
+		public int id {
 			get {
 				return _id;
 			}
